Move query execution code to query-runner module
