+++ java/athena/src/main/java/sleeper/athena/record/IteratorApplyingRecordHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Crown Copyright
+ * Copyright 2022-2024 Crown Copyright
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -51,8 +51,8 @@ import sleeper.core.schema.type.IntType;
 import sleeper.core.schema.type.LongType;
 import sleeper.core.schema.type.StringType;
 import sleeper.core.schema.type.Type;
-import sleeper.query.recordretrieval.LeafPartitionRecordRetrieverImpl;
-import sleeper.query.recordretrieval.RecordRetrievalException;
+import sleeper.query.runner.recordretrieval.LeafPartitionRecordRetrieverImpl;
+import sleeper.query.runner.recordretrieval.RecordRetrievalException;
 
 import java.io.Serializable;
 import java.util.ArrayList;
