+++ java/query/query-lambda/src/main/java/sleeper/query/lambda/SqsQueryProcessorLambda.java
@@ -31,6 +31,7 @@ import sleeper.configuration.jars.ObjectFactoryException;
 import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.configuration.properties.table.TablePropertiesProvider;
 import sleeper.core.util.LoggedDuration;
+import sleeper.query.runner.recordretrieval.QueryExecutor;
 import sleeper.query.runner.tracker.DynamoDBQueryTracker;
 
 import java.time.Instant;
@@ -41,7 +42,7 @@ import static sleeper.configuration.properties.instance.QueryProperty.QUERY_PROC
 
 /**
  * A lambda that is triggered when a serialised query arrives on an SQS queue. A processor executes the request using a
- * {@link sleeper.query.executor.QueryExecutor} and publishes the results to either SQS or S3 based on the configuration of the query.
+ * {@link QueryExecutor} and publishes the results to either SQS or S3 based on the configuration of the query.
  * The processor contains a cache that includes mappings from partitions to files in those partitions. This is reused by
  * subsequent calls to the lambda if the AWS runtime chooses to reuse the instance.
  */
