+++ java/query/query-runner/src/main/java/sleeper/query/runner/recordretrieval/LeafPartitionQueryExecutor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.query.recordretrieval;
+package sleeper.query.runner.recordretrieval;
 
 import org.apache.hadoop.conf.Configuration;
 import org.slf4j.Logger;
