+++ java/query/query-runner/src/main/java/sleeper/query/runner/recordretrieval/LeafPartitionRecordRetriever.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Crown Copyright
+ * Copyright 2022-2024 Crown Copyright
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.query.recordretrieval;
+package sleeper.query.runner.recordretrieval;
 
 import sleeper.core.iterator.CloseableIterator;
 import sleeper.core.record.Record;
