+++ java/query/query-runner/src/main/java/sleeper/query/runner/recordretrieval/RecordRetrievalException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Crown Copyright
+ * Copyright 2022-2024 Crown Copyright
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,17 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.query.recordretrieval;
+package sleeper.query.runner.recordretrieval;
 
 public class RecordRetrievalException extends Exception {
-
-    public RecordRetrievalException() {
-        super();
-    }
-
-    public RecordRetrievalException(String msg) {
-        super(msg);
-    }
 
     public RecordRetrievalException(String msg, Throwable cause) {
         super(msg, cause);
